get rid of speculate, pull few deps up to the workspace manifest
