+++ Cargo.toml
@@ -9,3 +9,11 @@ package.license = "GPL-3.0-only"
 
 [profile.release]
 lto = "fat"
+
+[workspace.dependencies]
+env_logger = "0.11"
+log = "0.4"
+thiserror = "1.0"
+
+stegano-seasmoke = { path = "crates/stegano-seasmoke", version = "0.1" }
+stegano-core = { path = "crates/stegano-core", version = "0.6.0" }
