+++ crates/stegano-core/Cargo.toml
@@ -23,23 +23,23 @@ codecov = { repository = "steganogram/stegano-rs", branch = "main", service = "g
 maintenance = { status = "passively-maintained" }
 
 [dependencies]
-image = { version = "0.24", default-features = false, features = [
+image = { version = "0.25", default-features = false, features = [
     "png",
     "jpeg",
 ] }
 bitstream-io = { version = "2.5" }
-zip = { version = "2.2.0", default-features = false, features = ["deflate"] }
+zip = { version = "2.2", default-features = false, features = ["deflate"] }
 byteorder = "1.5"
 hound = "3.5"
-thiserror = "1.0"
+thiserror.workspace = true
 enum_dispatch = "0.3"
-env_logger = "0.11"
-log = "0.4"
 
-stegano-seasmoke = { version = "0.1", path = "../stegano-seasmoke" }
+log.workspace = true
+env_logger.workspace = true
+
+stegano-seasmoke.workspace = true
 
 [dev-dependencies]
-speculate = "0.1"
 criterion = { version = "0.5", features = ["html_reports"] }
 tempfile = "3.12"
 coverage-helper = "0.2"
