+++ crates/stegano-seasmoke/Cargo.toml
@@ -14,7 +14,7 @@ argon2 = "0.5"
 rand = "0.8"
 chacha20poly1305 = { version = "0.10" }
 zeroize = "1.8"
-thiserror = "1.0"
+thiserror.workspace = true
 
 [build-dependencies]
 csbindgen = "1.9"
